fix: two lint failures on the retention sweep — a redundant cast and a redundant skip
`unconvert` names four `int64(n)` conversions in `runRetentionSweepWindow`.
All three store methods already return `int64`, so every one of them is a
no-op; they came from writing the assignments before checking the signatures.

The conditional-skip guard names the two endpoint tests. Their
`if testing.Short() { t.Skip(...) }` is redundant rather than wrong:
`testutil.SuiteTestDB` opens with exactly that check, and then does the part
that matters — an unreachable database is `Fatalf` when a DSN was configured
and `Skipf` only when none was. Duplicating the short-mode half at the call
site adds a skip site the guard has to account for and buys nothing the helper
was not already doing.

Verified against a live PostgreSQL: all six retention tests pass, including
both endpoint tests. `scripts/check-skips.sh` exits 0. `golangci-lint` with
`unconvert` is clean, and — the control — putting one conversion back makes it
exit 1 naming that line, so the lint run can distinguish the two trees.

+++ cmd/cleat-worker/setup.go
@@ -2593,12 +2593,12 @@ func (w *Worker) runRetentionSweepWindow(retentionDays, completedWorkflowRetenti
 		if n, err := w.store.DeleteExpiredEvents(w.ctx, cutoff); err != nil {
 			res.Errors = append(res.Errors, "events: "+err.Error())
 		} else {
-			res.EventsDeleted = int64(n)
+			res.EventsDeleted = n
 		}
 		if n, err := w.store.ClearExpiredCompactionState(w.ctx, cutoff); err != nil {
 			res.Errors = append(res.Errors, "compaction_state: "+err.Error())
 		} else {
-			res.CompactionStateCleared = int64(n)
+			res.CompactionStateCleared = n
 		}
 	} else {
 		res.Skipped = append(res.Skipped, "events and compaction_state (--retention-days is 0)")
@@ -2607,7 +2607,7 @@ func (w *Worker) runRetentionSweepWindow(retentionDays, completedWorkflowRetenti
 		if n, err := w.store.DeleteCompletedWorkflows(w.ctx, at(completedWorkflowRetentionDays)); err != nil {
 			res.Errors = append(res.Errors, "completed_workflows: "+err.Error())
 		} else {
-			res.CompletedWorkflows = int64(n)
+			res.CompletedWorkflows = n
 		}
 	} else {
 		res.Skipped = append(res.Skipped, "completed_workflows (--completed-workflow-retention-days is 0)")
@@ -2616,7 +2616,7 @@ func (w *Worker) runRetentionSweepWindow(retentionDays, completedWorkflowRetenti
 		if n, err := w.store.DeleteDeadLetteredWorkflows(w.ctx, at(deadLetterRetentionDays)); err != nil {
 			res.Errors = append(res.Errors, "dead_lettered: "+err.Error())
 		} else {
-			res.DeadLetteredWorkflows = int64(n)
+			res.DeadLetteredWorkflows = n
 		}
 	} else {
 		res.Skipped = append(res.Skipped, "dead_lettered (--dead-letter-retention-days is 0)")
